:recycle: refactor: 맵 마커 컴포넌트로 분리
